pyproject.toml
[project]
name = "domeok"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "greenlet>=3.3.2",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uuid6>=2025.0.1",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 120
